Add comprehensive unit tests for user override endpoints
Consolidates override tests into a single test file and adds comprehensive
test coverage for stage direction overrides to match the cue colour override
test suite. Uses parameterized subtests for validation error scenarios.

Test Coverage:
- GET endpoint (with and without overrides)
- POST endpoint (create overrides with full validation)
- POST validation errors (5 subtests for missing/invalid fields)
- POST invalid style ID handling
- PATCH endpoint (update overrides)
- PATCH permission checks (prevent unauthorized updates)
- DELETE endpoint (remove overrides)
- DELETE permission checks (prevent unauthorized deletions)
- CASCADE DELETE (verify orphaned overrides are cleaned up)

Key Changes:
- Moved cue colour override tests from separate file into test_overrides.py
- Added 8 new test methods for stage direction overrides (15 total test cases)
- Used subTest for parameterized validation tests (works with unittest.TestCase)
- Fixed setUp to create complete overrides with all required fields
- All tests verify database state changes and response codes
